fix(aci): add validation to automation edit view
